x86: jailhouse: Improve APIC support
This both fixes the x2APIC mode by setting the required access handlers
early (no idea yet why Linux isn't doing this in case it detects a BIOS-
enabled x2APIC), and it registers the APIC properly so that the correct
physical ID is known. The latter fixes MSI routing.