x86: jailhouse: Fix x2apic detection
The x2APIC is available if and only if it is already enabled.

This fixes the case that the root cell and, thus, also Jailhouse are
running in xAPIC mode while the hardware would support x2APIC. Even then
we need to stay in xAPIC mode with the non-root cell as Jailhouse
doesn't support switching modes.