spi: tegra114: move gpio initialisation to device probe()
devm_gpio_request must not be done in the setup() routine, it must be
done in the probe() routine, as the setup routine might be called more
often. Otherwise, subsequent calls to setup() will fail with EBUSY.

!!
This one still contains a bug!

Device unitialisation in case of errors in not done properly.
!!

Signed-off-by: Ralf Ramsauer <ralf@ramses-pyramidenbau.de>