arm: Initialize hrtimer-based broadcast clockevent
Analogously to 9358d755bd5c, this registers a broadcast clockevent in
case no hardware broadcast timer is available and the per-CPU timers can
be stopped in deep power states.

Partitions of the Jailhouse hypervisor fall in this category.
Registering the workaround timer allows to enter high-resolution mode in
that case.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>