plic: Add delay in plic_toggle
I don't understand it. With Jailhouse on the Rocket core, we need this
strange delay. Both calls, the readl and write to the PLIC should trap
the hypervisor. It occurs, that in most times, we stuck here at the
writel().

It seems like the writel() won't trap the hypervisor, instead, it holds
the whole platform. Adding a delay (or even other things, like a printk)
help to avoid it - for what reason ever.

It seems like there needs to be "some time" between two traps to the
hypervisor on the Rocketcore, but I'm uncertain about that.